packaging: Explicit add data files to the manifest
We've started seeing some failures to build in Debian and Ubuntu that
can be traced to those files missing from the manifest when we're
calling setup.py install.

Now, this started failing fairly recently, so I'm guessing it's actually
a change of behaviour in the SCM plugin and/or setuptools. It makes
sense though, since we're packaging from the tarball, which isn't in the
Git context needed to actually include those data files.

Signed-off-by: Simon Chopin <simon.chopin@canonical.com>