msgpack: increase max_buffer_size to ~4GiB
as borg now uses repository.store_load and .store_save to load
and save the chunks cache, we need a rather high limit here.

this is a quick fix, the real fix might be using chunks of the
data (preferably <= MAX_OBJECT_SIZE), so there is less to unpack
at once.