docs: update install docs, nothing bundled anymore! fixes #8342.
borg 1.2.x used to bundle some source code for lz4, zstd and xxhash,
but borg 1.4.x got rid of that - all that is meanwhile available as
a library in all OSes we support.

also, the python package pkgconfig does not need to be "preinstalled"
anymore, because our pyproject.toml cares for that. otoh, the cli tool
pkg-config must be preinstalled so that libs and headers can be found
automagically.

Also be a bit more clear about the FUSE stuff.