vagrant: pyenv: only use Python 3.11.11
Installing 3.9/10/12 involves downloading and compiling these,
which takes a lot of time and has some potential breaking the
boxes by installing misc. openssl versions.

We test misc. python versions on github actions CI all the time,
so we only need the one here which we want to use for the binary
builds.