blake2b_256: fix key passing, fixes #8867
borg 1.x involved the key a bit unusually in the blake2b
hash calculation (padding 64B of key material with another
64B of zeros to a total of 128B and then just prepending
it to the data).

since a while, we use blake2b from python stdlib and they
support passing the key as a separate argument. up to 64B
keys are allowed here.

besides the way how the 64B of key material are passed, there
are also other differences in how the key gets involved in
the digest computation, thus the legacy blake2b hashes from
borg 1.x are not compatible with the non-legacy ones - the
hash is not the same for same key and data values.