vagrant: drop macOS 10.12 VM
The VM was used for local macOS testing and
also for building a macOS intel fat binary.

We also do macOS CI testing on GitHub and I
recently added binary building on GitHub for
Apple Silicon and Intel.

The macOS 10 VM was very outdated, super slow
and a pain to use. I didn't succeed in building
a recent macOS vagrant VM, so we'll just use
GitHub from now on...