freebsd/netbsd: skip test_hard_link_deletion_and_replacement, #9147, #9153
The test fails on these platforms.

I could not find the root cause of this issue, but it is likely a minor
problem with ctime and doesn't affect borg usage much.

So I rather like to have CI on freebsd/netbsd not failing because of this.

Also: add is_netbsd and is_openbsd to platformflags.