archive: preserve pre-open original atime, fixes #6194
When creating an archive with --atime on platforms
without O_NOATIME, opening a file for reading could
update atime before we recorded it (thus we archived
the updated atime, not the original one).

Capture pre-open path-based-stat atime and use it
if it pre-dates the atime we got from the fd AND
if it refers to same fs object (avoid race condition).